dev (memory): separating broker memory managment from tree managment
