+++ cmd/insprd/memory/fake/mem_manager.go
@@ -1,7 +1,6 @@
 package fake
 
 import (
-	"github.com/inspr/inspr/pkg/meta/brokers"
 	"github.com/inspr/inspr/pkg/meta/utils/diff"
 
 	"github.com/inspr/inspr/cmd/insprd/memory"
@@ -15,7 +14,6 @@ type MemManager struct {
 	channel   Channels
 	app       Apps
 	alias     Alias
-	brokers   Brokers
 }
 
 // LookupMemManager mocks getter for roots
@@ -60,10 +58,6 @@ func MockMemoryManager(failErr error) memory.Manager {
 			fail:  failErr,
 			alias: make(map[string]*meta.Alias),
 		},
-		brokers: Brokers{
-			fail:   failErr,
-			broker: &brokers.Brokers{},
-		},
 	}
 }
 
@@ -104,9 +98,4 @@ func (mm *MemManager) Cancel() {}
 //GetTransactionChanges mock interface structure
 func (mm *MemManager) GetTransactionChanges() (diff.Changelog, error) {
 	return diff.Changelog{}, nil
-}
-
-//Brokers returns mocked manager interface
-func (mm *MemManager) Brokers() memory.BrokerInterface {
-	return &mm.brokers
 }
