+++ cmd/insprd/memory/interface.go
@@ -4,7 +4,6 @@ package memory
 
 import (
 	"github.com/inspr/inspr/pkg/meta"
-	"github.com/inspr/inspr/pkg/meta/brokers"
 	"github.com/inspr/inspr/pkg/meta/utils/diff"
 )
 
@@ -81,7 +80,6 @@ type Manager interface {
 	Types() TypeMemory
 	Alias() AliasMemory
 	Root() GetInterface
-	Brokers() BrokerInterface
 }
 
 // GetInterface is an interface to get components from memory
@@ -98,13 +96,4 @@ type TransactionInterface interface {
 	GetTransactionChanges() (diff.Changelog, error)
 	InitTransaction()
 	Cancel()
-}
-
-// BrokerInterface is the interface tht allows for interaction
-// with the systems multiple brokers
-type BrokerInterface interface {
-	GetAll() brokers.BrokerStatusArray
-	GetDefault() brokers.BrokerStatus
-	Create(broker brokers.BrokerStatus, config brokers.BrokerConfiguration) error
-	SetDefault(broker brokers.BrokerStatus) error
 }
